Bug fix: change "byte" to "int" to allow running more than 128 workers.

+++ gps-rev-110/src/java/gps/node/worker/dynamic/VertexWrapper.java
@@ -7,5 +7,5 @@ public class VertexWrapper<V extends MinaWritable> {
 	public int[] neighborIds;
 	public V state;
 	public boolean isActive;
-	public byte toOrFromMachineId;
+	public int toOrFromMachineId;
 }
